ref(tiller): move test only code into test file
