+++ cmd/tiller/kind_sorter.go
@@ -16,10 +16,6 @@ limitations under the License.
 
 package main
 
-import (
-	"sort"
-)
-
 // SortOrder is an ordering of Kinds.
 type SortOrder []string
 
@@ -28,48 +24,3 @@ var InstallOrder SortOrder = []string{"Namespace", "Secret", "ConfigMap", "Persi
 
 // UninstallOrder is the order in which manifests should be uninstalled (by Kind)
 var UninstallOrder SortOrder = []string{"Service", "Pod", "ReplicationController", "Deployment", "DaemonSet", "ConfigMap", "Secret", "PersistentVolume", "ServiceAccount", "Ingress", "Job", "Namespace"}
-
-// sortByKind does an in-place sort of manifests by Kind.
-//
-// Results are sorted by 'ordering'
-func sortByKind(manifests []manifest, ordering SortOrder) []manifest {
-	ks := newKindSorter(manifests, ordering)
-	sort.Sort(ks)
-	return ks.manifests
-}
-
-type kindSorter struct {
-	ordering  map[string]int
-	manifests []manifest
-}
-
-func newKindSorter(m []manifest, s SortOrder) *kindSorter {
-	o := make(map[string]int, len(s))
-	for v, k := range s {
-		o[k] = v
-	}
-
-	return &kindSorter{
-		manifests: m,
-		ordering:  o,
-	}
-}
-
-func (k *kindSorter) Len() int { return len(k.manifests) }
-
-func (k *kindSorter) Swap(i, j int) { k.manifests[i], k.manifests[j] = k.manifests[j], k.manifests[i] }
-
-func (k *kindSorter) Less(i, j int) bool {
-	a := k.manifests[i]
-	b := k.manifests[j]
-	first, ok := k.ordering[a.head.Kind]
-	if !ok {
-		// Unknown is always last
-		return false
-	}
-	second, ok := k.ordering[b.head.Kind]
-	if !ok {
-		return true
-	}
-	return first < second
-}
